fix: date picker ref type
