+++ components/date-picker/__tests__/type.test.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
+import { Moment } from 'moment';
 import DatePicker from '..';
-import type { DatePickerProps } from '..';
+import { DatePickRef, RangePickerRef } from '../generatePicker/interface';
 
 describe('DatePicker.typescript', () => {
   it('DatePicker ref methods', () => {
@@ -15,14 +16,10 @@ describe('DatePicker.typescript', () => {
     expect(datePicker).toBeTruthy();
   });
 
-
   // https://github.com/ant-design/ant-design/issues/33417
   it('DatePicker ref methods with forwardRef', () => {
-    const MyDatePicker = React.forwardRef<typeof DatePicker, DatePickerProps>((props, ref) => (
-      <DatePicker
-        {...props}
-        ref={ref}
-      />
+    const MyDatePicker = React.forwardRef((props, ref: DatePickRef<Moment>) => (
+      <DatePicker {...props} ref={ref} />
     ));
     const datePicker = (
       <MyDatePicker
@@ -45,5 +42,20 @@ describe('DatePicker.typescript', () => {
       />
     );
     expect(rangePicker).toBeTruthy();
+  });
+
+  it('RangePicker ref methods with forwardRef', () => {
+    const MyRangePicker = React.forwardRef((props, ref: RangePickerRef<Moment>) => (
+      <DatePicker.RangePicker {...props} ref={ref} />
+    ));
+    const datePicker = (
+      <MyRangePicker
+        ref={picker => {
+          picker?.focus();
+          picker?.blur();
+        }}
+      />
+    );
+    expect(datePicker).toBeTruthy();
   });
 });
