+++ components/date-picker/generatePicker/interface.ts
@@ -1,0 +1,83 @@
+import type { ComponentClass, Component, ForwardedRef } from 'react';
+import type { Locale as RcPickerLocale } from 'rc-picker/lib/interface';
+import type {
+  PickerBaseProps as RCPickerBaseProps,
+  PickerDateProps as RCPickerDateProps,
+  PickerTimeProps as RCPickerTimeProps,
+} from 'rc-picker/lib/Picker';
+import type { SizeType } from '../../config-provider/SizeContext';
+import type { TimePickerLocale } from '../../time-picker';
+import type {
+  RangePickerBaseProps as RCRangePickerBaseProps,
+  RangePickerDateProps as RCRangePickerDateProps,
+  RangePickerTimeProps as RCRangePickerTimeProps,
+} from 'rc-picker/lib/RangePicker';
+
+type InjectDefaultProps<Props> = Omit<
+  Props,
+  'locale' | 'generateConfig' | 'hideHeader' | 'components'
+> & {
+  locale?: PickerLocale;
+  size?: SizeType;
+  bordered?: boolean;
+};
+
+export type PickerLocale = {
+  lang: RcPickerLocale & AdditionalPickerLocaleLangProps;
+  timePickerLocale: TimePickerLocale;
+} & AdditionalPickerLocaleProps;
+
+export type AdditionalPickerLocaleProps = {
+  dateFormat?: string;
+  dateTimeFormat?: string;
+  weekFormat?: string;
+  monthFormat?: string;
+};
+
+export type AdditionalPickerLocaleLangProps = {
+  placeholder: string;
+  yearPlaceholder?: string;
+  quarterPlaceholder?: string;
+  monthPlaceholder?: string;
+  weekPlaceholder?: string;
+  rangeYearPlaceholder?: [string, string];
+  rangeMonthPlaceholder?: [string, string];
+  rangeWeekPlaceholder?: [string, string];
+  rangePlaceholder?: [string, string];
+};
+
+// Picker Props
+export type PickerBaseProps<DateType> = InjectDefaultProps<RCPickerBaseProps<DateType>>;
+export type PickerDateProps<DateType> = InjectDefaultProps<RCPickerDateProps<DateType>>;
+export type PickerTimeProps<DateType> = InjectDefaultProps<RCPickerTimeProps<DateType>>;
+
+export type PickerProps<DateType> =
+  | PickerBaseProps<DateType>
+  | PickerDateProps<DateType>
+  | PickerTimeProps<DateType>;
+
+// Range Picker Props
+export type RangePickerBaseProps<DateType> = InjectDefaultProps<RCRangePickerBaseProps<DateType>>;
+export type RangePickerDateProps<DateType> = InjectDefaultProps<RCRangePickerDateProps<DateType>>;
+export type RangePickerTimeProps<DateType> = InjectDefaultProps<RCRangePickerTimeProps<DateType>>;
+
+export type RangePickerProps<DateType> =
+  | RangePickerBaseProps<DateType>
+  | RangePickerDateProps<DateType>
+  | RangePickerTimeProps<DateType>;
+
+export interface CommonPickerMethods {
+  focus: () => void;
+  blur: () => void;
+}
+
+export interface PickerComponentClass<P = {}, S = unknown> extends ComponentClass<P, S> {
+  new (...args: ConstructorParameters<ComponentClass<P, S>>): InstanceType<ComponentClass<P, S>> &
+    CommonPickerMethods;
+}
+
+export type PickerRef<P> = ForwardedRef<Component<P> & CommonPickerMethods>;
+
+export type DatePickRef<DateType> = PickerRef<PickerProps<DateType>>;
+
+export type RangePickerRef<DateType> = PickerRef<RangePickerProps<DateType>>;
